feature : added onQueryStarted function to implement optimistic update scenario on follow button click.
