+++ src/modules/users/store/usersApi.ts
@@ -1,8 +1,48 @@
+import { RootState, baseAPI, AppDispatch } from '@/redux';
 import { ApiResponseTemplate } from '@/shared/types/response-data';
 import { UserId, UsersQueryParams, UsersResponse } from '@/modules/users/types';
 import { USERS, FOLLOW } from '@/shared/constants/query-endpoints';
 import { DELETE, POST } from '@/shared/constants/query-methods';
-import baseAPI from '@/redux/base-API';
+
+interface UpdateFollowStatusCacheParams {
+  state: RootState;
+  dispatch: AppDispatch;
+  usersApi: typeof usersApi;
+  userId: UserId;
+}
+
+const updateFollowStatusCache = ({
+  state,
+  dispatch,
+  usersApi,
+  userId,
+}: UpdateFollowStatusCacheParams) => {
+  const { queries } = state.api;
+  const getUsersQuery = Object.values(queries).find((query) => {
+    return query?.endpointName === 'getUsers';
+  });
+
+  if (!getUsersQuery?.originalArgs) {
+    throw new Error('Failed to find the query arguments for getUsers');
+  }
+
+  const { usersQueryCount, currentPage } =
+    getUsersQuery.originalArgs as UsersQueryParams;
+
+  return dispatch(
+    usersApi.util.updateQueryData<'getUsers'>(
+      'getUsers',
+      { usersQueryCount, currentPage },
+      (draft: UsersResponse) => {
+        const user = draft.items.find((user) => user.id === userId);
+
+        if (user) {
+          user.followed = !user.followed;
+        }
+      }
+    )
+  );
+};
 
 const usersApi = baseAPI.injectEndpoints({
   endpoints: (builder) => ({
@@ -17,17 +57,21 @@ const usersApi = baseAPI.injectEndpoints({
         method: DELETE,
         url: `${FOLLOW}/${userId}`,
       }),
-      // async onQueryStarted(userId, { dispatch, getState, queryFulfilled }) {
-      // await updateFollowStatusInCache({
-      //   userId,
-      //   dispatch,
-      //   getState,
-      //   queryFulfilled,
-      //   usersApi,
-      //   isFollowed: true,
-      // });
-      // },
-      invalidatesTags: ['Users'],
+
+      async onQueryStarted(userId, { dispatch, getState, queryFulfilled }) {
+        const patchResult = updateFollowStatusCache({
+          state: getState() as RootState,
+          dispatch,
+          usersApi,
+          userId,
+        });
+
+        try {
+          await queryFulfilled;
+        } catch (error) {
+          patchResult.undo();
+        }
+      },
     }),
 
     followUser: builder.mutation<ApiResponseTemplate, UserId>({
@@ -36,17 +80,20 @@ const usersApi = baseAPI.injectEndpoints({
         url: `${FOLLOW}/${userId}`,
       }),
 
-      // async onQueryStarted(userId, { dispatch, getState, queryFulfilled }) {
-      // await updateFollowStatusInCache({
-      //   userId,
-      //   dispatch,
-      //   getState,
-      //   queryFulfilled,
-      //   usersApi,
-      //   isFollowed: false,
-      // });
-      // },
-      invalidatesTags: ['Users'],
+      async onQueryStarted(userId, { dispatch, getState, queryFulfilled }) {
+        const patchResult = updateFollowStatusCache({
+          state: getState() as RootState,
+          dispatch,
+          usersApi,
+          userId,
+        });
+
+        try {
+          await queryFulfilled;
+        } catch (error) {
+          patchResult.undo();
+        }
+      },
     }),
   }),
 });
@@ -56,3 +103,5 @@ export const {
   useUnfollowUserMutation,
   useFollowUserMutation,
 } = usersApi;
+
+// export default usersApi;
